requirements.txt
requests==2.31.0
python-dateutil==2.8.2
numpy==1.24.3
python-kucoin>=2.1.3

